maybe make closure better *again*.
this delegates dealing with closing to the receiving task, instead of
the pumper internals. also allows receiving msgs during closure,
but not sending them.

the context manager ``__aexit__`` now directly eats all of the
messages from the channel in case there's a silly task still
trying to read, to avoid conflicts.